Generate all moves and rank
There's no lookahead, but at least we should attack when fighting a
monster and match keys when at a door or chest.

There seems to be a bug in lock detection where we switch back to thinking
attacks are demanded when at a door or chest, possibly due to match
graphic effects?

The ranking functions are kind of arbitrary and could be tuned.